Add the ability to output multiple Functions to a vtk file.
Also allow the user to specify a Lagrange p1, p2, p1dg or p2dg functionspace on which the vtkfile should be based.  This uses a
different (write) interface and allows GenericFunctions to be written to file.  A mesh based (rather than functionspace) interface is also provided.

The python interface and parallel have not been tested.  In particular there is currently no swig interface to allow outputting of
multiple functions (in a std::vector) and a time (as a double) in a std::pair so I have commented out the C++ versions of these
subroutines temporarily.