Fixing errors with vtk output in parallel when using a functionspace.
Most of these were bugs relating to local or global counts or indices to the dofmap.  Unfortunately the current solution is to do a
lot of evaluating of the dofs in the dofmap (including ghost nodes).  This could be cached somewhere to reduce cost (particularly of
sorting and eliminating duplicates from the dof vector).

Additionally there was a bug in pvtu output where the wrong xml node was being picked up when it was reopened (for multiple fields,
which had never before been tested).